+++ process/sync/shardblock.go
@@ -46,7 +46,7 @@ func NewShardBootstrap(
 	resolversFinder dataRetriever.ResolversFinder,
 	shardCoordinator sharding.Coordinator,
 	accounts state.AccountsAdapter,
-	boostrapRoundIndex uint32,
+	boostrapRoundIndex uint64,
 ) (*ShardBootstrap, error) {
 
 	if poolsHolder == nil {
